==> projects/exeves/stable/random_locations/01_crop_data.R <==
source('source/exeves.R')

region <- 'random_locations'

masks <- readRDS(paste0(PATH_SAVE, "partition_evap/evap_masks.rds"))
masks[elev_class == '0-100', .(lon, lat), KG_class_2]

set.seed(1)
random_kg_locations <- masks[masks[
  elev_class == '0-100' & 
    lat > 0 &
    rel_dataset_agreement != 'low' &
    rel_dataset_agreement != 'below average' &
    rel_dataset_agreement != 'average', 
  .I[sample(.N, 1)], by = KG_class_2]$V1]

random_kg_locations <- random_kg_locations[complete.cases(random_kg_locations)]

dataset_brick_evap <- brick(paste0(PATH_EVAP_SIM, 'gleam-v3-7a_e_mm_land_198001_202212_025_daily.nc'))

layer_n <- nlayers(dataset_brick_evap)

no_cores <- detectCores() - 1
if (no_cores < 1 | is.na(no_cores))(no_cores <- 1)
registerDoParallel(cores = no_cores)
dummie <- foreach (layer_count = 1:layer_n, .combine = rbind) %dopar% {
  dummie_layer <- dataset_brick_evap[[layer_count]]
  dummie_layer <- extract(dummie_layer, random_kg_locations[, .(lon, lat)])
  dummie_layer
}
colnames(dummie) <- random_kg_locations$KG_class_2
evap_kg_locations <- data.table(dummie)
evap_kg_locations[, date := dataset_brick_evap@z$Date]
dataset_dt <- melt(evap_kg_locations, id.vars = 'date')
colnames(dataset_dt)[2] <- 'KG_class_2'       
saveRDS(dataset_dt, paste0(PATH_OUTPUT_DATA, 'evap_', region, '_gleam.rds'))

plot(x = random_kg_locations$lon, y = random_kg_locations$lat)
